[utils] Extract utils from SakefileDescription
